refactor: migrate flow_classifier.py → shared/classifiers.py
- Create src/vibe3/services/shared/classifiers.py with flow classification logic
- Update both __init__.py files to add lazy import mappings
- Update flow_ui.py to import from vibe3.services instead of flow_classifier
- Fix status_query_service.py to import IssueCollectionService from shared.collections
- Fix all files importing from branch_arg to use vibe3.services imports
- Fix all files importing LabelService to use shared.labels import
- Delete old src/vibe3/services/flow_classifier.py

This migration consolidates flow classification utilities into the shared module
and ensures consistent import patterns across the codebase.

Note: Pre-existing mypy errors in spec_ref_service.py, flow_ui_primitives.py,
and handoff_storage.py are unrelated to this migration.
