+++ src/vibe3/services/blocked_state_io.py
@@ -17,8 +17,8 @@ from vibe3.clients.github_client import GitHubClient
 from vibe3.models import IssueState
 from vibe3.models.issue_body import FlowStateProjection
 from vibe3.services.blocked_state_types import BlockedState
-from vibe3.services.issue_body_service import merge_projection, parse_projection
-from vibe3.services.label_service import LabelService
+from vibe3.services.shared.issue_utils import merge_projection, parse_projection
+from vibe3.services.shared.labels import LabelService
 
 if TYPE_CHECKING:
     from vibe3.clients import SQLiteClient
