+++ src/vibe3/services/coordination_resolver.py
@@ -172,7 +172,7 @@ class CoordinationResolver:
             from body, or None if read failed
         """
         from vibe3.clients.github_client import GitHubClient
-        from vibe3.services.issue_body_service import parse_projection
+        from vibe3.services.shared.issue_utils import parse_projection
 
         client = GitHubClient()
