+++ src/vibe3/services/flow_recovery_service.py
@@ -263,7 +263,7 @@ class FlowRecoveryService:
         from vibe3.config.orchestra_settings import load_orchestra_config
         from vibe3.models import IssueInfo, IssueState
         from vibe3.services.flow_rebuild_usecase import FlowRebuildUsecase
-        from vibe3.services.issue_context_loader import load_issue_info
+        from vibe3.services.shared.issue_utils import load_issue_info
 
         # Load issue info from GitHub
         config = load_orchestra_config()
