+++ src/vibe3/services/flow_status_resolver.py
@@ -9,7 +9,7 @@ from loguru import logger
 from vibe3.models.data_source import DataSource
 from vibe3.models.flow import FlowStatusResponse
 from vibe3.services.flow_service import FlowService
-from vibe3.services.issue_body_service import parse_projection
+from vibe3.services.shared.issue_utils import parse_projection
 
 if TYPE_CHECKING:
     from vibe3.clients import SQLiteClient
